# docs/exercise-api-docs.json
{
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "https://exercisedb-api-mauve.vercel.app",
            "description": "v1 Dataset (Open Source)\n• Public & open license\n• Code and metadata available on GitHub\n• GIF-based media\n• Ideal for demos, personal apps, and learning\n• chat support for full dataset access"
        }
    ],
...
    "paths": {
...
        "/api/v1/exercises": {
            "get": {
                "tags": [
                    "EXERCISES"
                ],
                "summary": "Get all exercises with optional search",
                "description": "Retrieve all exercises with optional fuzzy search filtering",
                "operationId": "getExercises",
                "parameters": [
                    {
                        "schema": {
                            "type": "number",
                            "description": "The number of exercises to skip from the start of the list. Useful for pagination to fetch subsequent pages of results.",
                            "title": "Offset",
                            "example": 0,
                            "default": 0
                        },
                        "required": false,
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "number",
                            "description": "The maximum number of exercises to return in the response. Limits the number of results for pagination purposes.",
                            "title": "Limit",
                            "maximum": 25,
                            "minimum": 1,
                            "example": 10,
                            "default": 10
                        },
                        "required": false,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "description": "Optional search term for fuzzy matching across all exercise fields",
                            "title": "Search Query",
                            "example": "cardio",
                            "default": ""
                        },
                        "required": false,
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "exerciseId",
                                "targetMuscles",
                                "bodyParts",
                                "equipments"
                            ],
                            "default": "targetMuscles",
                            "description": "Field to sort exercises by",
                            "title": "Sort Field",
                            "example": "targetMuscles"
                        },
                        "required": false,
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "default": "desc",
                            "description": "Sort order (ascending or descending)",
                            "title": "Sort Order",
                            "example": "desc"
                        },
                        "required": false,
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with exercises",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ],
                                            "description": "Indicates whether the request was successful",
                                            "example": true
                                        },
                                        "metadata": {
                                            "type": "object",
                                            "properties": {
                                                "totalExercises": {
                                                    "type": "number",
                                                    "description": "Total number of exercises matching the criteria",
                                                    "example": 150
                                                },
                                                "totalPages": {
                                                    "type": "number",
                                                    "description": "Total number of pages available",
                                                    "example": 15
                                                },
                                                "currentPage": {
                                                    "type": "number",
                                                    "description": "Current page number",
                                                    "example": 1
                                                },
                                                "previousPage": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "URL for the previous page, null if on first page",
                                                    "example": "/api/exercises?offset=0&limit=10"
                                                },
                                                "nextPage": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "URL for the next page, null if on last page",
                                                    "example": "/api/exercises?offset=20&limit=10"
                                                }
                                            },
                                            "required": [
                                                "totalExercises",
                                                "totalPages",
                                                "currentPage",
                                                "previousPage",
                                                "nextPage"
                                            ]
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "exerciseId": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "gifUrl": {
                                                        "type": "string"
                                                    },
                                                    "targetMuscles": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "bodyParts": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "equipments": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "secondaryMuscles": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "instructions": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "required": [
                                                    "exerciseId",
                                                    "name",
                                                    "gifUrl",
                                                    "targetMuscles",
                                                    "bodyParts",
                                                    "equipments",
                                                    "secondaryMuscles",
                                                    "instructions"
                                                ]
                                            },
                                            "description": "Array of exercises"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "metadata",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/v1/exercises/{exerciseId}": {
            "get": {
                "tags": [
                    "EXERCISES"
                ],
                "summary": "GetExerciseById",
                "operationId": "getExerciseById",
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "description": "The unique identifier of the exercise to retrieve.",
                            "title": "Exercise ID",
                            "example": "ztAa1RK",
                            "default": "ztAa1RK"
                        },
                        "required": true,
                        "name": "exerciseId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the exercise details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "description": "Indicates whether the request was successful.",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "exerciseId": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "gifUrl": {
                                                    "type": "string"
                                                },
                                                "targetMuscles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "bodyParts": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "equipments": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "secondaryMuscles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "instructions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "required": [
                                                "exerciseId",
                                                "name",
                                                "gifUrl",
                                                "targetMuscles",
                                                "bodyParts",
                                                "equipments",
                                                "secondaryMuscles",
                                                "instructions"
                                            ],
                                            "description": "The retrieved exercise details."
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exercise not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}